requirements.txt
django
djangorestframework
drf-yasg
psycopg2-binary
django-extensions
gunicorn
django-cors-headers